# web/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jeevanti Hospitals Management System API",
    "description": "Comprehensive API documentation for the Jeevanti Hospitals Management System (HMS) built with Next.js, Prisma, MongoDB, and NextAuth.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/invoices": {
      "get": {
        "summary": "List invoices",
        "description": "Retrieves a list of invoices.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Invoice"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      },
      "post": {
        "summary": "Create invoice",
        "description": "Creates a custom invoice with a set of line items.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
    "/api/invoices/auto-opd": {
      "post": {
        "summary": "Auto-generate OPD invoice",
        "description": "Auto-generates an invoice for an OPD consultation based on consultation fees, prescribed medicines, and ordered lab tests.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "consultationId"
                ],
                "properties": {
                  "consultationId": {
                    "type": "string",
                    "description": "OPD Consultation ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OPD Invoice auto-generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
    "/api/invoices/auto-ipd": {
      "post": {
        "summary": "Auto-generate IPD invoice",
        "description": "Auto-generates/updates invoice for inpatient stay including room charges, lab tests, and additional services.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "admissionId"
                ],
                "properties": {
                  "admissionId": {
                    "type": "string",
                    "description": "IPD Admission ID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "IPD Invoice auto-generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "500": {
            "$ref": "#/components/responses/500InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "context": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "admissionId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "OPD",
              "IPD"
            ]
          },
          "subtotal": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "partial",
              "paid",
              "cancelled"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "paidAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "unitPrice": {
                  "type": "number"
                },
                "total": {
                  "type": "number"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "method": {
                  "type": "string"
                },
                "transactionId": {
                  "type": "string"
                },
                "paidAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "receivedBy": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "InvoiceInput": {
        "type": "object",
        "required": [
          "patientId",
          "type",
          "items"
        ],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "admissionId": {
            "type": "string"
          },
          "appointmentId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "OPD",
              "IPD"
            ]
          },
          "tax": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "description",
                "quantity",
                "unitPrice",
                "type"
              ],
              "properties": {
                "description": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "unitPrice": {
                  "type": "number"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "400ValidationError": {
        "description": "Validation failed / Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "404NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "500InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}